[feat][admin] Unify findFirstLedgerWithinThreshold logic in cli and admin
