+++ pulsar-client-admin/src/test/java/org/apache/pulsar/client/admin/internal/PulsarAdminImplTest.java
@@ -18,10 +18,15 @@
  */
 package org.apache.pulsar.client.admin.internal;
 
+import java.util.ArrayList;
+import java.util.List;
 import lombok.SneakyThrows;
 import org.apache.pulsar.client.api.Authentication;
+import org.apache.pulsar.client.impl.MessageIdImpl;
 import org.apache.pulsar.client.impl.auth.AuthenticationDisabled;
 import org.apache.pulsar.client.impl.conf.ClientConfigurationData;
+import org.apache.pulsar.common.policies.data.ManagedLedgerInternalStats;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -44,6 +49,35 @@ public class PulsarAdminImplTest {
         ClientConfigurationData conf = new ClientConfigurationData();
         conf.setAuthentication(auth);
         assertThat(createAdminAndGetAuth(conf)).isSameAs(auth);
+    }
+
+    @Test
+    public void testFindFirstLedgerWithinThreshold() throws Exception {
+        List<ManagedLedgerInternalStats.LedgerInfo> ledgers = new ArrayList<>();
+        ledgers.add(newLedger(0, 10, 1000));
+        ledgers.add(newLedger(1, 10, 2000));
+        ledgers.add(newLedger(2, 10, 3000));
+
+        // test huge threshold
+        Assert.assertNull(TopicsImpl.findFirstLedgerWithinThreshold(ledgers, Long.MAX_VALUE));
+
+        // test small threshold
+        Assert.assertEquals(TopicsImpl.findFirstLedgerWithinThreshold(ledgers, 0),
+                new MessageIdImpl(2, 0, -1));
+
+        // test middling thresholds
+        Assert.assertEquals(TopicsImpl.findFirstLedgerWithinThreshold(ledgers, 1000),
+                new MessageIdImpl(2, 0, -1));
+        Assert.assertEquals(TopicsImpl.findFirstLedgerWithinThreshold(ledgers, 5000),
+                new MessageIdImpl(1, 0, -1));
+    }
+
+    private static ManagedLedgerInternalStats.LedgerInfo newLedger(long id, long entries, long size) {
+        ManagedLedgerInternalStats.LedgerInfo l = new ManagedLedgerInternalStats.LedgerInfo();
+        l.ledgerId = id;
+        l.entries = entries;
+        l.size = size;
+        return l;
     }
 
     @SneakyThrows
